+++ node.gyp
@@ -846,24 +846,24 @@
         [ 'node_use_openssl=="true" and experimental_quic==1', {
           'defines': ['NODE_EXPERIMENTAL_QUIC=1'],
           'sources': [
-            'src/node_quic_buffer.h',
-            'src/node_quic_crypto.h',
-            'src/node_quic_session.h',
-            'src/node_quic_session-inl.h',
-            'src/node_quic_socket.h',
-            'src/node_quic_stream.h',
-            'src/node_quic_util.h',
-            'src/node_quic_util-inl.h',
-            'src/node_quic_state.h',
-            'src/node_quic_default_application.h',
-            'src/node_quic_http3_application.h',
-            'src/node_quic_crypto.cc',
-            'src/node_quic_session.cc',
-            'src/node_quic_socket.cc',
-            'src/node_quic_stream.cc',
-            'src/node_quic.cc',
-            'src/node_quic_default_application.cc',
-            'src/node_quic_http3_application.cc'
+            'src/quic/node_quic_buffer.h',
+            'src/quic/node_quic_crypto.h',
+            'src/quic/node_quic_session.h',
+            'src/quic/node_quic_session-inl.h',
+            'src/quic/node_quic_socket.h',
+            'src/quic/node_quic_stream.h',
+            'src/quic/node_quic_util.h',
+            'src/quic/node_quic_util-inl.h',
+            'src/quic/node_quic_state.h',
+            'src/quic/node_quic_default_application.h',
+            'src/quic/node_quic_http3_application.h',
+            'src/quic/node_quic_crypto.cc',
+            'src/quic/node_quic_session.cc',
+            'src/quic/node_quic_socket.cc',
+            'src/quic/node_quic_stream.cc',
+            'src/quic/node_quic.cc',
+            'src/quic/node_quic_default_application.cc',
+            'src/quic/node_quic_http3_application.cc'
           ]
         }
         ],
