+++ test/cctest/test_quic_buffer.cc
@@ -1,4 +1,4 @@
-#include "node_quic_buffer.h"
+#include "quic/node_quic_buffer.h"
 #include "util-inl.h"
 #include "uv.h"
